+++ src/main/java/hackerton/caudex/domain/garden/dto/GardenResDto.java
@@ -41,7 +41,7 @@ public class GardenResDto {
             Long plantId,
             Double ratioX,
             Double ratioY,
-            Double scale,
+            Integer scale,
             String plantUrl
     ){}
